Install security and upload guard plugins at startup
